refactor: home submit detail の構成規則を統一
