+++ src/features/home/sections/TopPlayersPanel.tsx
@@ -1,7 +1,7 @@
 ﻿import type { Bracket } from "../../../data/mockRuns";
 import { HOME_BUTTON_CLASSES, HOME_LABELS, HOME_PANEL_CLASSES, HOME_TOP_PLAYER_BRACKETS } from "../homeConfig";
 import { getBestRun } from "../homeLogic";
-import { TopPlayerCard } from "../TopPlayerCard";
+import { TopPlayerCard } from "../ui/TopPlayerCard";
 import type { HomeRun } from "../types";
 
 type TopPlayersPanelProps = {
