+++ src/features/recordDetail/recordDetailLogic.ts
@@ -1,0 +1,1 @@
+﻿export * from "./logic/recordDetailHelpers";
